% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_tickers.R
\name{default_tickers}
\alias{default_tickers}
\title{Default stock tickers available in the app}
\usage{
default_tickers
}
\description{
A character vector of pharma company tickers used as app defaults.
}
