% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_operators.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x, y}{Any R objects.}
}
\value{
\code{x} if \code{!is.null(x)}, else \code{y}.
}
\description{
Returns \code{x} if it is non-\code{NULL}, otherwise returns \code{y}.  A lightweight
alternative to \verb{rlang::\%||\%} that avoids adding a heavy dependency.
}
\examples{
NULL \%||\% "default"   # "default"
"value" \%||\% "other"  # "value"

}
