% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_tickers.R
\docType{data}
\name{default_tickers}
\alias{default_tickers}
\title{Default stock tickers available in the app}
\format{
An object of class \code{character} of length 7.
}
\usage{
default_tickers
}
\description{
A character vector of pharma company tickers used as app defaults.
}
\keyword{datasets}
